* Moved workspace implementation into folder * Removed old generate templates * Build command works * Started on publish command (incomplete)
