+++ package.json
@@ -9,7 +9,7 @@
     "test": "NODE_ENV=test mocha --exit"
   },
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=14.0.0"
   },
   "repository": {
     "type": "git",
@@ -33,9 +33,11 @@
     "@hapi/joi": "17.1.1",
     "boxen": "5.1.2",
     "chalk": "4.1.2",
+    "cli-progress": "3.9.1",
     "cli-table3": "0.6.0",
     "commander": "8.2.0",
     "debug": "4.3.2",
+    "dockerode": "3.3.1",
     "find-package-json": "1.2.0",
     "form-data": "4.0.0",
     "get-stream": "6.0.1",
