--- src/commands/assets/templates/skill/daemon/Dockerfile
@@ -1,5 +1,0 @@
-FROM python:3.7.4-buster
-ADD . /app
-WORKDIR /app
-RUN pip install -r requirements.txt
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "5000", "--workers", "2"]
