--- src/commands/assets/templates/skill/daemon/action.json.hbs
@@ -1,8 +1,0 @@
-{
-    "name":"{{dashCase name}}",
-    "actionType":"daemon",
-    "docker":"{{dashCase name}}:latest",
-    "environmentVariables":"\"TEST\"=\"value\"",
-    "command": ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "5000", "--workers", "2"],
-    "port": "5000"
-}
