--- src/commands/assets/templates/skill/daemon/fabric.yaml
@@ -1,7 +1,0 @@
-version: 1
-kind: deployment-manifest
-cortex:
-  skills:
-    - skill.yaml
-  actions:
-    - action.json
