--- src/commands/assets/templates/skill/daemon/main.py.hbs
@@ -1,9 +1,0 @@
-from fastapi import FastAPI
-
-app = FastAPI()
-
-
-@app.post('/invoke')
-def run(requestBody: dict):
-    payload = requestBody['payload']
-    return {'payload': payload}
