--- src/commands/assets/templates/skill/daemon/requirements.txt.hbs
@@ -1,3 +1,0 @@
-cortex-python
-fastapi
-uvicorn
