--- src/commands/assets/templates/skill/daemon/skill.yaml.hbs
@@ -1,46 +1,0 @@
-camel: 1.0.0
-name: {{dashCase name}}
-title: {{name}}
-description: {{name}}
-inputs:
-  - name: request
-    title: API Request
-    parameters:
-      - name: params
-        type: object
-        description: Request Parameters
-        required: true
-    routing:
-      all:
-        action  : {{dashCase name}}
-        runtime: cortex/daemons
-        output: response
-
-outputs:
-  - name: response
-    title: API Response
-    parameters:
-      - name: result
-        type: object
-        description: API Response
-        required: true
-
-properties:
-  - name: daemon.method
-    title: Daemon method
-    description: Update default value to HTTP method supported by endpoint
-    required: true
-    type: String
-    defaultValue: POST
-  - name: daemon.path
-    title: Daemon path
-    description: Update default value to HTTP endpoint path in container
-    required: true
-    type: String
-    defaultValue: invoke
-  - name: daemon.port
-    title: Daemon Port
-    description: Update default value to port on which app will be running
-    required: true
-    type: String
-    defaultValue: 5000
