--- src/commands/assets/templates/skill/daemon/test/payload.json
@@ -1,3 +1,0 @@
-{
-  "payload": {}
-}
