--- src/commands/assets/templates/skill/external-api/test/payload.json
@@ -1,3 +1,0 @@
-{
-  "payload": {}
-}
