--- src/commands/assets/templates/skill/job/action.json.hbs
@@ -1,7 +1,0 @@
-{
-    "name":"{{dashCase name}}",
-    "actionType":"job",
-    "docker":"{{dashCase name}}:latest",
-    "environmentVariables":"\"TEST\"=\"value\"",
-    "command": ["python", "/app/main.py"]
-}
