--- src/commands/assets/templates/skill/job/main.py.hbs
@@ -1,7 +1,0 @@
-import json
-
-if __name__ == '__main__':
-    import sys
-    params = sys.argv[1]
-    params = json.loads(params)
-    print(f'Received: {params}')
