--- src/commands/assets/templates/skill/job/requirements.txt.hbs
@@ -1,1 +1,0 @@
-cortex-python
