--- src/commands/assets/templates/skill/job/skill.yaml.hbs
@@ -1,26 +1,0 @@
-camel: 1.0.0
-name: {{dashCase name}}
-title: {{name}}
-description: {{name}}
-inputs:
-  - name: params
-    title: Job Input
-    parameters:
-      - name: params
-        type: object
-        description: Input Params
-        required: true
-    routing:
-      all:
-        action  : {{dashCase name}}
-        runtime: cortex/jobs
-        output: result
-
-outputs:
-  - name: result
-    title: Job result
-    parameters:
-      - name: result
-        type: object
-        description: Job Execution Logs
-        required: true
