--- src/commands/assets/templates/skill/job/test/payload.json
@@ -1,3 +1,0 @@
-{
-  "payload": {}
-}
